feat: transitioned to a cli app. can start the project in different mode using the two main commands.
refactor: minor refactoring around borrowing some values for the functions

chore: removed the unused rrl_core and config_loader project that are now merged into rate_limiter.

chore: the project will be a binary, the configuration file won't be coming with it so an example will live at the root of the project.